libbio: add casts to remove -Wconversion warnings
Update #5764

R=golang-dev, dave, rsc
CC=golang-dev
https://golang.org/cl/12388043

+++ src/libbio/bgetc.c
@@ -49,7 +49,7 @@ loop:
 	 * buffer to allow that many ungets.
 	 */
 	memmove(bp->bbuf-Bungetsize, bp->ebuf-Bungetsize, Bungetsize);
-	i = read(bp->fid, bp->bbuf, bp->bsize);
+	i = (int)read(bp->fid, bp->bbuf, (size_t)bp->bsize);
 	bp->gbuf = bp->bbuf;
 	if(i <= 0) {
 		bp->state = Bracteof;
@@ -58,7 +58,7 @@ loop:
 		return Beof;
 	}
 	if(i < bp->bsize) {
-		memmove(bp->ebuf-i-Bungetsize, bp->bbuf-Bungetsize, i+Bungetsize);
+		memmove(bp->ebuf-i-Bungetsize, bp->bbuf-Bungetsize, (size_t)(i+Bungetsize));
 		bp->gbuf = bp->ebuf-i;
 	}
 	bp->icount = -i;
